requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
idna==3.4
install==1.3.5
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytesseract==0.3.10
regex==2023.5.4
requests==2.29.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.65.0
urllib3==1.26.15
